massive refactor not finished)
