+++ bioimage_phenotyping/features/importance.py
@@ -1,5 +1,36 @@
 import pandas as pd
 import shap
+from scipy import stats
+import seaborn as sns
+import pandas as pd
+import matplotlib.pyplot as plt
+import numpy as np
+from sklearn.preprocessing import StandardScaler
+from sklearn.preprocessing import scale
+from sklearn.decomposition import PCA
+from sklearn.metrics.pairwise import euclidean_distances
+from sklearn.ensemble import IsolationForest
+from sklearn.preprocessing import scale, power_transform, robust_scale, normalize
+from sklearn import model_selection
+from sklearn.ensemble import RandomForestClassifier, BaggingClassifier
+from sklearn import metrics
+from sklearn.metrics import classification_report
+from sklearn.pipeline import Pipeline
+from sklearn.feature_selection import SelectFromModel, RFECV, RFE
+
+import functools
+import shap
+from sklearn.ensemble import RandomForestClassifier
+from sklearn.model_selection import train_test_split
+from sklearn.preprocessing import LabelEncoder
+from sklearn.pipeline import Pipeline
+from sklearn.preprocessing import StandardScaler, PowerTransformer
+from sklearn.base import BaseEstimator
+from sklearn.base import TransformerMixin
+
+
+import os
+
 
 # TODO, typee hinting
 def get_shap_df(shaps):
@@ -9,9 +40,10 @@ def get_shap_df(shaps):
         .reset_index()
         .melt(id_vars="Sample", var_name="Feature", value_name="Shap Value")
     )
-    
+
+
 def get_shap_values(
-    self,
+    df,
     model,
     variable="Cell",
     groupby=None,
@@ -21,7 +53,7 @@ def get_shap_values(
     *args,
     **kwargs,
 ):
-    df = self.df
+
     # X = np.array(df.apply(pd.to_numeric))
     X, y = df, list(df.index.get_level_values(variable))
     X100 = shap.utils.sample(np.array(X), 100)
@@ -29,7 +61,7 @@ def get_shap_values(
     # y = df.index.get_level_values(variable)
     # y = DistanceMatrix().fit_transform(X)
 
-    X_train, X_test, y_train, y_test = self.df.apply(
+    X_train, X_test, y_train, y_test = df.apply(
         pd.to_numeric
     ).bip.train_test_split(variable, groupby=groupby, augment=augment)
     # model = RandomForestClassifier()
@@ -47,3 +79,43 @@ def get_shap_values(
 
     shap_values = explainer(X)
     return shap_values
+
+
+def feature_importances(
+    df,
+    model_class=RandomForestClassifier,
+    variable="Cell",
+    groupby=None,
+    augment=None,
+    kfolds=1,
+):
+    importance_list = []
+
+    for fold in range(1, kfolds + 1):
+        model = model_class()
+        X_train, X_test, y_train, y_test = (
+            self.df
+            # .balance_dataset(variable)
+            .bip.train_test_split(variable, groupby=groupby, augment=augment, seed=fold)
+        )
+        model.fit(X_train.values, y_train)
+
+        print(classification_report(y_test, model.predict(X_test)))
+        print(metrics.cohen_kappa_score(y_test, model.predict(X_test)))
+
+        importance = (
+            pd.DataFrame(
+                model.feature_importances_,
+                index=pd.Series(X_train.columns, name="Feature"),
+                columns=["Importance"],
+            )
+            .assign(Fold=fold)
+            .sort_values(ascending=False, by="Importance")
+        )
+
+        importance["Cumulative Importance"] = importance.cumsum()["Importance"]
+        importance.attrs.update(self.df.attrs)
+        importance_list.append(importance)
+        # scores = self.get_score_df_from_model(model,variable,X_test,y_test)
+
+    return pd.concat(importance_list)
